requirements.txt
flask
pandas
google-cloud-bigquery
scikit-surprise
db-dtypes
gunicorn